refactor: Build handleSendAnimalResponseSms helper
