+++ helpers/index.js
@@ -1,4 +1,9 @@
+import handleSendAnimalResponseSms from './handleSendAnimalResponseSms';
 import handleSendContinentsResponseSms from './handleSendContinentsResponseSms';
 import handleSendInitialSms from './handleSendInitialSms';
 
-export { handleSendContinentsResponseSms, handleSendInitialSms };
+export {
+  handleSendAnimalResponseSms,
+  handleSendContinentsResponseSms,
+  handleSendInitialSms,
+};
